Use Histogram objects
This is a pretty large commit (whoops) that changes the way projections
and diagnostics are defined in MENT. We already had the Histogram1D and
HistogramND classes that were used as diagnostics. These objects store
the coordinates of the histogram bin centers/edges and act as functions
that bin particles on the grid. I've now added a `values` parameter to
these classes to store the bin counts.

Now a "projection" and "diagnostic" are both an instance of HistogramND
or Histogram1D. The only thing the user needs to do is create a list
of histograms called `projections` from the measured profiles and pass
this to MENT. MENT will then make a copy of this list and call it
`diagnostics`, which previously was passed by the user. The `diagnostics`
list has the same histogram bins but is updated when MENT simulates the
projections; `projections` is never changed.

I also updated the plotting routines to use `ultraplot`, a fork of `proplot`
that is being actively developed. I also use the latest version of `psdist`.
The plotting routines in this repo are included for convenience and are
stored in the `ment.train` module.

All example jupyter notebook have been saved with full output to see the
reconstruction results on GitHub.