Add integration_loop parameter to MENT
MENT needs to compute the projection array g. Assume we flatten the
array, so it is a one-dimensional array of value g[i]. Then we
need to compute an integral for each i. Right now we loop over
each i and compute the integral.

If integration_loop=False, we instead evaluate the MENT PDF on the
entire meshgrid of projection_grid times integration_grid. This is
an N-dimensional grid. This speeds up integration, so is useful for
2D problems, or possibly 4D problems, but is probably infeasible for
6D problems on personal computers.